タイトルシーンの描画追加

+++ DX9LibX64/Class/SceneManager/Scene/Scene.h
@@ -28,5 +28,5 @@ protected:
 	InputData m_inputData;
 	BaseWorkingManager m_sceneObjectManager;
 	FileManager m_fileManager;
-	INT m_frame = -1;
+	LONGLONG m_frame = -1;
 };
